+++ dynamic_routing_gateways.tf
@@ -60,5 +60,5 @@ resource "oci_core_drg" "these" {
   #Optional
   defined_tags  = each.value.defined_tags
   display_name  = each.value.display_name
-  freeform_tags = each.value.freeform_tags
+  freeform_tags = merge(local.cislz_module_tag, each.value.freeform_tags)
 }