+++ service_gateways.tf
@@ -3,11 +3,12 @@
 # Licensed under the Universal Permissive License v 1.0 as shown at https: //oss.oracle.com/licenses/upl. #
 # Author: Cosmin Tudor                                                                                    #
 # Author email: cosmin.tudor@oracle.com                                                                   #
-# Last Modified: Wed Nov 15 2023                                                                          #
+# Last Modified: Tue Dec 19 2023                                                                          #
 # Modified by: Cosmin Tudor, email: cosmin.tudor@oracle.com                                               #
 # ####################################################################################################### #
 
 data "oci_core_services" "oci_services" {
+  count = var.network_configuration != null ? 1 : 0
 }
 
 locals {
@@ -16,7 +17,7 @@ locals {
   //   - all-services
   //   - objectstorage
   oci_services_details = {
-    for oci_service in data.oci_core_services.oci_services.services : length(regexall("services-in-oracle-services-network", oci_service.cidr_block)) > 0 ? "all-services" : length(regexall("objectstorage", oci_service.cidr_block)) > 0 ? "objectstorage" : oci_service.cidr_block => oci_service
+    for oci_service in length(data.oci_core_services.oci_services) > 0 ? data.oci_core_services.oci_services[0].services : [] : length(regexall("services-in-oracle-services-network", oci_service.cidr_block)) > 0 ? "all-services" : length(regexall("objectstorage", oci_service.cidr_block)) > 0 ? "objectstorage" : oci_service.cidr_block => oci_service
   }
 
   one_dimension_processed_service_gateways = local.one_dimension_processed_vcn_specific_gateways != null ? {
@@ -120,7 +121,7 @@ resource "oci_core_service_gateway" "these" {
   #Optional
   defined_tags  = each.value.defined_tags
   display_name  = each.value.display_name
-  freeform_tags = each.value.freeform_tags
+  freeform_tags = merge(local.cislz_module_tag, each.value.freeform_tags)
   // Searching for the id based on the key in the:
   //       - IGW and NAT GW specific route tables: local.provisioned_igw_natgw_specific_route_tables
   //       - SGW specific route tables: local.provisioned_sgw_specific_route_tables + the default route table